Refactor naming && try catch for not applicable file formats
