+++ MediaGalleryMetadata/Model/Gif/ReadFile.php
@@ -12,15 +12,16 @@ use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Filesystem\DriverInterface;
 use Magento\MediaGalleryMetadataApi\Model\FileInterface;
 use Magento\MediaGalleryMetadataApi\Model\FileInterfaceFactory;
-use Magento\MediaGalleryMetadataApi\Model\FileReaderInterface;
+use Magento\MediaGalleryMetadataApi\Model\ReadFileInterface;
 use Magento\MediaGalleryMetadataApi\Model\SegmentInterface;
 use Magento\MediaGalleryMetadataApi\Model\SegmentInterfaceFactory;
 use Magento\MediaGalleryMetadata\Model\SegmentNames;
+use Magento\Framework\Exception\ValidatorException;
 
 /**
  * File segments reader
  */
-class FileReader implements FileReaderInterface
+class ReadFile implements ReadFileInterface
 {
     /**
      * @var DriverInterface
@@ -61,35 +62,17 @@ class FileReader implements FileReaderInterface
     }
 
     /**
-     * Is the current file reader applicable to a given path
-     *
-     * @param string $path
-     * @return bool
-     */
-    private function isApplicable(string $path): bool
-    {
-        $resource = $this->driver->fileOpen($path, 'rb');
-        $marker = $this->read($resource, 3);
-        $this->driver->fileClose($resource);
-
-        return $marker == 'GIF';
-    }
-
-    /**
      * @inheritdoc
      */
     public function execute(string $path): ?FileInterface
     {
-        if (!$this->isApplicable($path)) {
-            return null;
-        }
         $resource = $this->driver->fileOpen($path, 'rb');
 
         $header = $this->read($resource, 3);
 
         if ($header != "GIF") {
             $this->driver->fileClose($resource);
-            throw new LocalizedException(__('Not a GIF image'));
+            throw new ValidatorException(__('Not an GIF image'));
         }
 
         $version = $this->read($resource, 3);
