+++ MediaGalleryMetadata/Model/Jpeg/ReadFile.php
@@ -12,15 +12,16 @@ use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Filesystem\DriverInterface;
 use Magento\MediaGalleryMetadataApi\Model\FileInterface;
 use Magento\MediaGalleryMetadataApi\Model\FileInterfaceFactory;
-use Magento\MediaGalleryMetadataApi\Model\FileReaderInterface;
+use Magento\MediaGalleryMetadataApi\Model\ReadFileInterface;
 use Magento\MediaGalleryMetadataApi\Model\SegmentInterface;
 use Magento\MediaGalleryMetadataApi\Model\SegmentInterfaceFactory;
 use Magento\MediaGalleryMetadata\Model\SegmentNames;
+use Magento\Framework\Exception\ValidatorException;
 
 /**
  * Jpeg file reader
  */
-class FileReader implements FileReaderInterface
+class ReadFile implements ReadFileInterface
 {
     private const MARKER_IMAGE_FILE_START = "\xD8";
     private const MARKER_PREFIX = "\xFF";
@@ -69,46 +70,21 @@ class FileReader implements FileReaderInterface
     }
 
     /**
-     * Is the current file reader applicable to a given path
-     *
-     * @param string $path
-     * @return bool
-     */
-    private function isApplicable(string $path): bool
-    {
-        $resource = $this->driver->fileOpen($path, 'rb');
-        try {
-            $marker = $this->readMarker($resource);
-        } catch (LocalizedException $exception) {
-            return false;
-        }
-        $this->driver->fileClose($resource);
-
-        return $marker == self::MARKER_IMAGE_FILE_START;
-    }
-
-    /**
      * @inheritdoc
      */
     public function execute(string $path): ?FileInterface
     {
         $segments = [];
-        
-        if (!$this->isApplicable($path)) {
-            return null;
-        }
-        
         $resource = $this->driver->fileOpen($path, 'rb');
-
-        $marker = $this->readMarker($resource);
+        $marker = $this->read($resource, self::TWO_BYTES)[1];
 
         if ($marker != self::MARKER_IMAGE_FILE_START) {
             $this->driver->fileClose($resource);
-            throw new LocalizedException(__('Not a JPEG image'));
+            throw new ValidatorException(__('Not a JPEG image'));
         }
 
         do {
-            $marker = $this->readMarker($resource);
+            $marker = $this->read($resource, self::TWO_BYTES)[1];
             $segments[] = $this->readSegment($resource, ord($marker));
         } while (($marker != self::MARKER_IMAGE_START) && (!$this->driver->endOfFile($resource)));
 
@@ -165,25 +141,6 @@ class FileReader implements FileReaderInterface
             'name' => $this->segmentNames->getSegmentName($segmentType),
             'data' => $this->read($resource, $segmentSize)
         ]);
-    }
-
-    /**
-     * Read jpeg marker
-     *
-     * @param resource $resource
-     * @return string
-     * @throws FileSystemException
-     */
-    private function readMarker($resource): string
-    {
-        $data = $this->read($resource, self::TWO_BYTES);
-
-        if ($data[0] != self::MARKER_PREFIX) {
-            $this->driver->fileClose($resource);
-            throw new LocalizedException(__('File is corrupted'));
-        }
-
-        return $data[1];
     }
 
     /**
