+++ MediaGalleryMetadata/Model/Png/Segment/WriteXmp.php
@@ -5,25 +5,24 @@
  */
 declare(strict_types=1);
 
-namespace Magento\MediaGalleryMetadata\Model\Jpeg\Segment;
+namespace Magento\MediaGalleryMetadata\Model\Png\Segment;
 
 use Magento\MediaGalleryMetadata\Model\AddXmpMetadata;
-use Magento\MediaGalleryMetadata\Model\XmpTemplate;
 use Magento\MediaGalleryMetadataApi\Api\Data\MetadataInterface;
 use Magento\MediaGalleryMetadataApi\Model\FileInterface;
 use Magento\MediaGalleryMetadataApi\Model\FileInterfaceFactory;
 use Magento\MediaGalleryMetadataApi\Model\MetadataWriterInterface;
 use Magento\MediaGalleryMetadataApi\Model\SegmentInterface;
 use Magento\MediaGalleryMetadataApi\Model\SegmentInterfaceFactory;
+use Magento\MediaGalleryMetadata\Model\XmpTemplate;
 
 /**
- * Jpeg XMP Writer
+ * XMP Writer for png format
  */
-class XmpWriter implements MetadataWriterInterface
+class WriteXmp implements MetadataWriterInterface
 {
-    private const XMP_SEGMENT_NAME = 'APP1';
-    private const XMP_SEGMENT_START = "http://ns.adobe.com/xap/1.0/\x00";
-    private const XMP_DATA_START_POSITION = 29;
+    private const XMP_SEGMENT_NAME = 'iTXt';
+    private const XMP_SEGMENT_START = "XML:com.adobe.xmp\x00";
 
     /**
      * @var SegmentInterfaceFactory
@@ -64,7 +63,7 @@ class XmpWriter implements MetadataWriterInterface
     }
 
     /**
-     * Add metadata to the file
+     * Add xmp metadata to the png file
      *
      * @param FileInterface $file
      * @param MetadataInterface $metadata
@@ -73,21 +72,21 @@ class XmpWriter implements MetadataWriterInterface
     public function execute(FileInterface $file, MetadataInterface $metadata): FileInterface
     {
         $segments = $file->getSegments();
-        $xmpSegments = [];
+        $pngXmpSegments = [];
         foreach ($segments as $key => $segment) {
-            if ($this->isSegmentXmp($segment)) {
-                $xmpSegments[$key] = $segment;
+            if ($this->isXmpSegment($segment)) {
+                $pngXmpSegments[$key] = $segment;
             }
         }
 
-        if (empty($xmpSegments)) {
+        if (empty($pngXmpSegments)) {
             return $this->fileFactory->create([
                 'path' => $file->getPath(),
-                'segments' => $this->insertXmpSegment($segments, $this->createXmpSegment($metadata))
+                'segments' => $this->insertPngXmpSegment($segments, $this->createPngXmpSegment($metadata))
             ]);
         }
 
-        foreach ($xmpSegments as $key => $segment) {
+        foreach ($pngXmpSegments as $key => $segment) {
             $segments[$key] = $this->updateSegment($segment, $metadata);
         }
 
@@ -98,24 +97,24 @@ class XmpWriter implements MetadataWriterInterface
     }
 
     /**
-     * Insert XMP segment to image segments (at position 1)
+     * Insert XMP segment to image png segments (at position 1)
      *
      * @param SegmentInterface[] $segments
      * @param SegmentInterface $xmpSegment
      * @return SegmentInterface[]
      */
-    private function insertXmpSegment(array $segments, SegmentInterface $xmpSegment): array
+    private function insertPngXmpSegment(array $segments, SegmentInterface $xmpSegment): array
     {
         return array_merge(array_slice($segments, 0, 2), [$xmpSegment], array_slice($segments, 2));
     }
 
     /**
-     * Write new segment  metadata
+     * Write new png segment  metadata
      *
      * @param MetadataInterface $metadata
      * @return SegmentInterface
      */
-    private function createXmpSegment(MetadataInterface $metadata): SegmentInterface
+    public function createPngXmpSegment(MetadataInterface $metadata): SegmentInterface
     {
         $xmpData = $this->xmpTemplate->get();
         return $this->segmentFactory->create([
@@ -125,7 +124,7 @@ class XmpWriter implements MetadataWriterInterface
     }
 
     /**
-     * Add metadata to the segment
+     * Add metadata to the png xmp segment
      *
      * @param SegmentInterface $segment
      * @param MetadataInterface $metadata
@@ -133,24 +132,32 @@ class XmpWriter implements MetadataWriterInterface
      */
     private function updateSegment(SegmentInterface $segment, MetadataInterface $metadata): SegmentInterface
     {
-        $data = $segment->getData();
-        $start = substr($data, 0, self::XMP_DATA_START_POSITION);
-        $xmpData = substr($data, self::XMP_DATA_START_POSITION);
         return $this->segmentFactory->create([
             'name' => $segment->getName(),
-            'data' => $start . $this->addXmpMetadata->execute($xmpData, $metadata)
+            'data' => self::XMP_SEGMENT_START . $this->addXmpMetadata->execute($this->getXmpData($segment), $metadata)
         ]);
     }
 
     /**
-     * Check if segment contains XMP data
+     * Does segment contain XMP data
      *
      * @param SegmentInterface $segment
      * @return bool
      */
-    private function isSegmentXmp(SegmentInterface $segment): bool
+    private function isXmpSegment(SegmentInterface $segment): bool
     {
         return $segment->getName() === self::XMP_SEGMENT_NAME
-            && strncmp($segment->getData(), self::XMP_SEGMENT_START, self::XMP_DATA_START_POSITION) == 0;
+            && strpos($segment->getData(), '<x:xmpmeta') !== -1;
+    }
+
+    /**
+     * Get XMP xml
+     *
+     * @param SegmentInterface $segment
+     * @return string
+     */
+    private function getXmpData(SegmentInterface $segment): string
+    {
+        return substr($segment->getData(), strpos($segment->getData(), '<x:xmpmeta'));
     }
 }
